Applied suggestions

+++ models/correlomatrix_detector.cpp
@@ -348,8 +348,8 @@ nest::correlomatrix_detector::handle( SpikeEvent& e )
       for ( SpikelistType::const_iterator spike_j = otherSpikes.begin(); spike_j != otherSpikes.end(); ++spike_j )
       {
         size_t bin;
-        long other = spike_j->receptor_channel_;
-        long sender_ind, other_ind;
+        size_t other = spike_j->receptor_channel_;
+        size_t sender_ind, other_ind;
 
         if ( spike_i < spike_j->timestep_ )
         {
@@ -378,14 +378,14 @@ nest::correlomatrix_detector::handle( SpikeEvent& e )
         {
           // weighted histogram
           S_.covariance_[ sender_ind ][ other_ind ][ bin ] += e.get_multiplicity() * e.get_weight() * spike_j->weight_;
-          if ( bin == 0 and ( spike_i - spike_j->timestep_ != 0 or static_cast< size_t >( other ) != sender ) )
+          if ( bin == 0 and ( spike_i - spike_j->timestep_ != 0 or other != sender ) )
           {
             S_.covariance_[ other_ind ][ sender_ind ][ bin ] +=
               e.get_multiplicity() * e.get_weight() * spike_j->weight_;
           }
           // pure (unweighted) count histogram
           S_.count_covariance_[ sender_ind ][ other_ind ][ bin ] += e.get_multiplicity();
-          if ( bin == 0 and ( spike_i - spike_j->timestep_ != 0 or static_cast< size_t >( other ) != sender ) )
+          if ( bin == 0 and ( spike_i - spike_j->timestep_ != 0 or other != sender ) )
           {
             S_.count_covariance_[ other_ind ][ sender_ind ][ bin ] += e.get_multiplicity();
           }
